fix:refactored project controls and add label, image models
