+++ labellab-server/routes/routes.js
@@ -4,10 +4,12 @@ var router = express.Router()
 // API's path
 var authRoute = require("./auth/routes")
 var usersRoute = require("./users/routes")
+var projectRoute = require("./project/routes")
 
 // Routes
 // -> /api/auth/
 router.use("/api/v1/auth", authRoute)
 router.use("/api/v1/users", usersRoute)
+router.use("/api/v1/project", projectRoute)
 
 module.exports = router
